grc_backend/requirements.txt
# ============================================================
# CORE DJANGO & REST
# ============================================================
Django>=4.2.7,<5.1
djangorestframework>=3.14.0,<3.16
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
django-filter>=23.4
django-extensions>=3.2.3
django-import-export>=3.3.4
django-simple-history>=3.4.0
drf-yasg>=1.21.7


# ============================================================
# AUTH / SECURITY
# ============================================================
PyJWT>=2.8.0
python-dotenv>=1.0.0
python-decouple>=3.8
cryptography>=41.0.7
bcrypt>=4.1.2


# ============================================================
# DATABASES
# ============================================================
mysqlclient>=2.2.0
PyMySQL>=1.1.0
mysql-connector-python>=8.0.33
SQLAlchemy>=2.0.23,<2.1


# ============================================================
# DATA
# ============================================================
numpy>=1.25.2,<2.0
pandas>=2.1.4,<2.2


# ============================================================
# DOCUMENT / PDF / OCR
# ============================================================
openpyxl>=3.1.2
xlsxwriter>=3.1.0
python-docx>=1.1.0
reportlab>=4.0.8
pdfplumber>=0.10.3
PyPDF2>=3.0.1
PyMuPDF>=1.23.8
xmltodict>=0.13.0
Pillow>=10.1.0
pytesseract>=0.3.10


# ============================================================
# ASYNC / TASKS
# ============================================================
requests>=2.31.0
aiohttp>=3.9.1
celery>=5.3.6
redis>=5.0.1
fakeredis>=2.23.2


# ============================================================
# AWS / STORAGE
# ============================================================
boto3>=1.34.0
django-storages>=1.14.2


# ============================================================
# LOGGING
# ============================================================
python-json-logger>=2.0.7
django-debug-toolbar>=4.2.0


# ============================================================
# AI / LANGCHAIN (FINAL FIX)
# ============================================================
openai>=1.40.0,<2.0.0

langchain>=0.2.17,<0.3.0
langchain-core>=0.2.43,<0.3.0
langchain-community>=0.2.16,<0.3.0
langchain-openai>=0.1.22,<0.2.0
langchain-ollama==0.1.3


# ============================================================
# VECTOR DB / EMBEDDINGS
# ============================================================
sentence-transformers>=2.2.2,<3.0
chromadb>=0.4.24,<0.5.0


# ============================================================
# VALIDATION
# ============================================================
pydantic>=2.7,<2.9
# Note: pydantic 2.8.2 is compatible with openai 1.75.0
# ollama requires pydantic>=2.9.0, but this causes compatibility issues with openai


# ============================================================
# PRODUCTION
# ============================================================
gunicorn>=21.2.0
whitenoise>=6.6.0
werkzeug>=2.3.8,<3.0


google-auth>=2.0.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

cerberus>=1.3.5

psycopg2

